+++ RELEASE_NOTES.md
@@ -1,17 +1,29 @@
+#### 0.0.6 - October 28 2017
+* Legivel.Mapper: map "null" for option types to "None"
+* Legivel.Mapper: Made "null" detection generic/injectable
+* Legivel.Mapper: removed field "StopLocation" in Succes result
+* Added documentation for Legivel customization
+* Added wishlist to documentation
+
+
 #### 0.0.5 - October 25 2017
 * Fixing too many framework dependencies issue
 * Having another shot at nuget release
+
 
 #### 0.0.4 - October 24 2017
 * Fixing Nuget dependencies
 * Adding more info to Nuge package
 
+
 #### 0.0.3 - October 23 2017
 * First Nuget package
+
 
 #### 0.0.2 - October 22 2017
 * Renamed product to 'Legivel'
 * Wrote majority of the documentation
+
 
 #### 0.0.1 - December 29 2016
 * Initial source publication on Github (started in sept 2016)